Marine Licence specific tests
